+++ PyPWA/builtin_plugins/gamp/g_iterator.py
@@ -17,12 +17,17 @@
 #  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 """
-Gamp data reading and writing.
+GAMP data reading and writing.
 
-This file holds the Gamp Reader and Gamp Writer. These simply load the
+This file holds the GAMP Reader and GAMP Writer. These simply load the
 data into memory one event at a time and write to file one event at a
 time. Only the previous loaded events are stored, anything later than that
 will not be saved in memory by these object.
+
+- GampReader: This reads in GAMP events from disk, GAMP events are deque
+  of named tuples, each named tuple representing a particle in the event.
+- GampWriter: Takes GAMP events one at a time and attempts to write them
+  in a standardized way to file so that other programs can read the output.
 """
 
 import io
@@ -32,13 +37,15 @@ import numpy
 from PyPWA import Path, AUTHOR, VERSION
 from PyPWA.libs import misc_file_libs
 from PyPWA.libs.components.data_processor import data_templates
+from PyPWA.libs.math import particle
+
 
 __credits__ = ["Mark Jones"]
 __author__ = AUTHOR
 __version__ = VERSION
 
 
-class GampParticleCount(object):
+class _GampParticleCount(object):
 
     def __init__(self):
         self.__particle_count = 0
@@ -62,26 +69,15 @@ class GampParticleCount(object):
 class GampReader(data_templates.Reader):
 
     def __init__(self, file_location):
-        """
-        This reads in Gamp events from disk, Gamp events are deque of
-        named tuples, each named tuple representing a particle in the
-        event.
-
-        Args:
-            file_location (Path): Name of the GAMP file, can be any size.
-        """
+        # type: (Path) -> None
         self.__event_count = None
-        self._previous_event = None  # type: numpy.ndarray
+        self._file = None  # type: io.BufferedReader
         self._the_file = file_location
-        self.__particle_count = GampParticleCount()
+        self.__particle_count = _GampParticleCount()
         self.__particle_count.get_particle_count(file_location)
         self._start_input()
 
     def _start_input(self):
-        """
-        Checks to see if the file is opened, and if it is close it so that
-        it may be reopened.
-        """
         try:
             if self._file:
                 self._file.close()
@@ -90,58 +86,35 @@ class GampReader(data_templates.Reader):
         self._file = self._the_file.open("rt")
 
     def reset(self):
-        """
-        Calls the _start_input method again which will close the file then
-        reopen it back at the beginning of the file.
-        """
         self._start_input()
 
     def next(self):
-        """
-        Structures the read in event from the GAMP file into a deque then
-        passes it to the calling function.
-
-        Returns:
-            deque: The next deque event.
-
-        Raises:
-            StopIterator: End of file has been found.
-        """
         first_line = self._file.readline().strip("\n")
         if first_line == "":
             raise StopIteration
-        count = int(first_line)
-        event = numpy.zeros((count, 6), numpy.float64)
-        for index in range(count):
-            event[index] = self._make_particle(self._file.readline())
-        self._previous_event = event
-        return self._previous_event
+
+        particle_count = int(first_line)
+        event = []
+        for index in range(particle_count):
+            event.append(self._make_particle(self._file.readline()))
+
+        return particle.ParticlePool(event)
 
     @staticmethod
     def _make_particle(string):
-        """
-        Takes the string read in from the GAMP event and parses it into a
-        named tuple to store the various values. All values are numpy data
-        types.
+        # type: (str) -> particle.Particle
 
-        Args:
-            string (str): The string containing the GAMP Particle
+        data_list = string.strip("\n").split(" ")
 
-        Returns:
-            numpy.ndarray: 2 dimensional array,
-                [particle_index][particles]
-        """
-        the_list = string.strip("\n").split(" ")
+        vector = numpy.zeros(1, particle.NUMPY_PARTICLE_DTYPE)
+        particle_id = numpy.float64(data_list[0])  # Particle ID
+        charge = numpy.float64(data_list[1])  # Particle Charge
+        vector['x'] = numpy.float64(data_list[2])  # Particle X Momentum
+        vector['y'] = numpy.float64(data_list[3])  # Particle Y Momentum
+        vector['z'] = numpy.float64(data_list[4])  # Particle Z Momentum
+        vector['e'] = numpy.float64(data_list[5])  # Particle Energy
 
-        particle = numpy.zeros(6, dtype=numpy.float64)
-        particle[0] = numpy.float64(the_list[0])  # Particle ID
-        particle[1] = numpy.float64(the_list[1])  # Particle Charge
-        particle[2] = numpy.float64(the_list[2])  # Particle X Momentum
-        particle[3] = numpy.float64(the_list[3])  # Particle Y Momentum
-        particle[4] = numpy.float64(the_list[4])  # Particle Z Momentum
-        particle[5] = numpy.float64(the_list[5])  # Particle Energy
-
-        return particle
+        return particle.Particle(particle_id, charge, vector)
 
     def get_event_count(self):
         return self.__particle_count.particle_count
@@ -153,28 +126,26 @@ class GampReader(data_templates.Reader):
 class GampWriter(data_templates.Writer):
 
     def __init__(self, file_location):
-        """
-        Takes GAMP events one at a time and attempts to write them in a
-        standardized way to file so that other programs can read the
-        output.
+        # type: (Path) -> None
+        self.__file = file_location.open("w")
 
-        Args:
-            file_location (Path): Where to write the GAMP data.
-        """
-        self._file = file_location.open("w")
+    def write(self, particle_pool):
+        # type: (particle.ParticlePool) -> None
+        self.__file.write(u"%d\n" % particle_pool.particle_count)
+        for the_particle in particle_pool.iterate_over_particles():
+            self.__write_particle(the_particle)
 
-    def write(self, data):
-        """
-        Writes the events the disk one event at a time, wont close the
-        disk access until the close function is called or until the object
-        is deleted.
-
-        Args:
-            numpy.ndarray: the file that is to be written to disk.
-        """
-        self._file.write(u"%d\n" % len(data))
-        for particle in data:
-            self._file.write(u"{0} {1} {2} {3} {4} {5}\n".format(*particle))
+    def __write_particle(self, the_particle):
+        # type: (particle.Particle) -> None
+        if len(the_particle) > 1:
+            raise ValueError("Can't write more than one event at a time!")
+        self.__file.write(
+            u'%s %s %s %s %s %s\n' % (
+                the_particle.id.astype(str), the_particle.charge.astype(str),
+                the_particle.x[0].astype(str), the_particle.y[0].astype(str),
+                the_particle.z[0].astype(str), the_particle.e[0].astype(str)
+            )
+        )
 
     def close(self):
-        self._file.close()
+        self.__file.close()