+++ PyPWA/builtin_plugins/numpy/__init__.py
@@ -58,5 +58,5 @@ class NumpyDataPlugin(data_templates.DataPlugin):
         return True
 
     @property
-    def plugin_supports_tree_data(self):
+    def plugin_supports_particle_pool(self):
         return False